Added middleware file

+++ app/api/issues/[id]/route.ts
@@ -1,50 +1,54 @@
+import authOptions from "@/app/auth/authOptions";
 import { issueSchema } from "@/app/validation";
 import prisma from "@/prisma/client";
+import { getServerSession } from "next-auth";
 import { NextRequest, NextResponse } from "next/server";
 
-export async function PATCH( 
-    request: NextRequest,
-    { params }: { params: { id: string }}
-){
-    const body = await request.json();
-    const validiation = issueSchema.safeParse(body);
-    if ( !validiation.success ) 
-        return NextResponse.json(validiation.error.format(), { status: 400 });
+export async function PATCH(
+  request: NextRequest,
+  { params }: { params: { id: string } }
+) {
+  const session = await getServerSession(authOptions);
+  if (!session) return NextResponse.json({}, { status: 401 });
+  const body = await request.json();
+  const validiation = issueSchema.safeParse(body);
+  if (!validiation.success)
+    return NextResponse.json(validiation.error.format(), { status: 400 });
 
-    const issue = await prisma.issue.findUnique({
-        where: { id: parseInt( params.id )}
-    })
+  const issue = await prisma.issue.findUnique({
+    where: { id: parseInt(params.id) },
+  });
 
-    if (!issue) 
-        return NextResponse.json({ error: 'Invalid Issue '}, { status: 404 })
+  if (!issue)
+    return NextResponse.json({ error: "Invalid Issue " }, { status: 404 });
 
-    const updatedIssue = await prisma.issue.update({
-        where: { id: issue.id},
-        data: {
-            title: body.title,
-            description: body.description
-        }
-    })
+  const updatedIssue = await prisma.issue.update({
+    where: { id: issue.id },
+    data: {
+      title: body.title,
+      description: body.description,
+    },
+  });
 
-    return NextResponse.json( updatedIssue )
+  return NextResponse.json(updatedIssue);
 }
 
+export async function DELETE(
+  request: NextRequest,
+  { params }: { params: { id: string } }
+) {
+  const session = await getServerSession(authOptions);
+  if (!session) return NextResponse.json({}, { status: 401 });
+  const issue = await prisma.issue.findUnique({
+    where: { id: parseInt(params.id) },
+  });
 
+  if (!issue)
+    return NextResponse.json({ error: "Invalid Issue " }, { status: 404 });
 
-export async function DELETE( 
-    request: NextRequest,
-    { params }: { params: { id: string }}
-){
-    const issue = await prisma.issue.findUnique({
-        where: { id: parseInt( params.id )}
-    })
+  const deletedIssue = await prisma.issue.delete({
+    where: { id: issue.id },
+  });
 
-    if (!issue) 
-        return NextResponse.json({ error: 'Invalid Issue '}, { status: 404 })
-
-    const deletedIssue = await prisma.issue.delete({
-        where: { id: issue.id}
-    })
-
-    return NextResponse.json( {} )
+  return NextResponse.json({});
 }
